+++ DangProject/Sources/Domain Layer/Protocols/FetchGraphDataUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  FetchGraphDataUseCase.swift
+//  DangProject
+//
+//  Created by 김동우 on 2022/08/06.
+//
+
+import Foundation
+
+protocol FetchGraphDataUseCase {
+    
+}
