+++ DangProject/Sources/Domain Layer/Protocols/Firebase/FirebaseFireStoreUseCase.swift
@@ -11,7 +11,6 @@ import RxSwift
 
 protocol ManageFirebaseFireStoreUseCase {
     var profileExistenceObservable: PublishSubject<Bool> { get }
-    var yearMonthDayDataSubject: PublishSubject<GraphData> { get }
     func uploadFirebaseUID(uid: String)
     func uploadProfile(profile: ProfileDomainModel)
     func getProfileExistence(uid: String) -> Observable<Bool>
@@ -19,6 +18,4 @@ protocol ManageFirebaseFireStoreUseCase {
     func getEatenFoods(dateComponents : DateComponents) -> Observable<[FoodDomainModel]>
     func getProfileData() -> Observable<ProfileDomainModel>
     func updateProfileData(_ data: ProfileDomainModel)
-    func createGraphThisYearMonthDayData()
-    func uploadDangAverage(_ data: Int)
 }
